Refactoring

+++ genres-service/src/main/java/com/javatar/genresservice/repository/GenreRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GenreRepository extends JpaRepository<Genre, Long> {
+public interface GenreRepository extends JpaRepository<Genre, Integer> {
     List<Genre> findByReviewid(Integer reviewid);
 }
