+++ labels-service/src/main/java/com/javatar/labelsservice/repository/LabelRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface LabelRepository extends JpaRepository<Label, Long> {
+public interface LabelRepository extends JpaRepository<Label, Integer> {
     List<Label> findByReviewid(Integer reviewid);
 }
