+++ years-service/src/main/java/com/javatar/yearsservice/repository/YearRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface YearRepository extends JpaRepository<Year, Long> {
+public interface YearRepository extends JpaRepository<Year, Integer> {
     List<Year> findByReviewid(Integer reviewid);
 }
